+++ docs/src/changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - External inflow to the [SBM + Local inertial river (1D) and land (2D)](@ref) model
   configuration.
+- Multiplication factors for precip and potential evaporation and additive factor for temperature to facilitate climate runs with delta change approach. 
 
 ## v0.6.1 - 2022-04-26